QPPCT-566: Move @AroundEach to new annotation package
